requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
python-dotenv
pydantic
pydantic-settings
langchain==0.3.26
langchain-google-genai==2.1.8
langgraph
httpx
google-auth
google-auth-oauthlib
google-api-python-client==2.176.0
tenacity
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
google-auth-oauthlib
google-auth-httplib2
pyyaml
